Renamed graphql/queries to .ts. Added prop types to search-comp
